django_service/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.4
Django==4.2
django-cors-headers==4.9.0
django-environ==0.12.0
django-restframework==0.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.11
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
PyMySQL==1.1.2
python-dotenv==1.2.2
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
